Refactor cookie management and remove ngx-cookie-service dependency
