+++ frontend/src/app/app.config.ts
@@ -4,11 +4,9 @@ import { provideRouter } from '@angular/router';
 import { routes } from './app.routes';
 import {provideHttpClient, withFetch} from '@angular/common/http';
 import { provideApollo } from 'apollo-angular';
-import { ApolloLink, InMemoryCache} from '@apollo/client/core';
-import {setContext} from "@apollo/client/link/context";
 import {environment} from "../environments/environment";
-import {HttpLink} from "apollo-angular/http";
 import {loadDevMessages, loadErrorMessages} from "@apollo/client/dev";
+import { createApolloClientOptions } from "./apollo.config";
 
 export const appConfig: ApplicationConfig = {
   providers: [
@@ -20,30 +18,8 @@ export const appConfig: ApplicationConfig = {
         loadDevMessages();
         loadErrorMessages();
       }
-      const httpLink = inject(HttpLink);
 
-      const authLink = setContext(() => {
-        const token = document.cookie
-          .split('; ')
-          .find(row => row.startsWith('token='))
-          ?.split('=')[1];
-        return {
-
-          headers: {
-            Authorization: token ? `Bearer ${token}` : '',
-          }
-        };
-      });
-
-      const link = ApolloLink.from([
-        authLink,
-        httpLink.create({ uri: environment.auth_service }),
-      ]);
-
-      return {
-        link: link,
-        cache: new InMemoryCache(),
-      };
+      return createApolloClientOptions();
     })
   ]
 };
