--- frontend/src/app/core/services/token.service.ts
@@ -1,23 +1,0 @@
-import { Injectable } from '@angular/core';
-import {CookieService} from "ngx-cookie-service";
-import {Token} from "../models/token.model";
-
-@Injectable({
-  providedIn: 'root'
-})
-export class TokenService {
-
-  constructor(private cookieService: CookieService) { }
-
-  get(): any {
-    this.cookieService.get("token");
-  }
-
-  set(token: Token): void {
-    this.cookieService.set("token", token.accessToken, token.expiresIn);
-  }
-
-  delete(): void {
-    this.cookieService.delete("token");
-  }
-}
