+++ frontend/src/app/features/auth/services/auth.service.ts
@@ -2,24 +2,16 @@ import {Injectable} from '@angular/core';
 import {Apollo} from "apollo-angular";
 import {map, Observable} from "rxjs";
 import {LOGIN, REGISTER} from "../requests/mutations";
-import {TokenService} from "../../../core/services/token.service";
 import {MutationResponse} from "../responses/mutation.response";
 import {QueryResponses} from "../responses/query.responses";
 import {IS_AUTHENTICATED} from "../requests/queries";
-import {environment} from "../../../../environments/environment";
-import {InMemoryCache} from "@apollo/client/core";
 
 @Injectable({
   providedIn: 'root'
 })
 export class AuthService {
 
-  constructor(private apollo: Apollo, private tokenService: TokenService) {
-    // apollo.create({
-    //   uri: environment.auth_service,
-    //   cache: new InMemoryCache()
-    // });
-  }
+  constructor(private apollo: Apollo) { }
 
   public login(username: string, password: string): Observable<boolean> {
     return this.apollo.mutate<MutationResponse>({
@@ -30,7 +22,6 @@ export class AuthService {
         const token = response.data?.login;
 
         if (token) {
-          this.tokenService.set(token);
           return true;
         } else {
           console.error("Login failed: no token received.");
@@ -55,7 +46,6 @@ export class AuthService {
         const token = response.data?.register;
 
         if (token) {
-          this.tokenService.set(token);
           return true;
         } else {
           console.error("Registration failed: no token received.");
